Clippy is now working

+++ cdrs_ci/src/lib.rs
@@ -1,4 +1,5 @@
 use std::collections::HashMap;
+use std::env::{current_dir, set_current_dir};
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
@@ -19,14 +20,23 @@ pub fn setup(workflow_dir: &Path) -> (Vec<String>, [(&'static str, &'static str)
 
     assert!(output.status.success());
 
-    // if !output.stdout.is_empty() {
-    //     panic!("Uncommitted changes, please commit them first: {:#?}", String::from_utf8(output.stdout).unwrap());
-    // }
+    if !output.stdout.is_empty() {
+        panic!(
+            "Uncommitted changes, please commit them first: {}",
+            String::from_utf8(output.stdout).unwrap()
+        );
+    }
 
     log::info!("Removing old workflow folders");
     // Ignore results, since maybe the folders do not exists atm
     let _ = std::fs::remove_dir_all(workflow_dir);
     let _ = std::fs::create_dir_all(workflow_dir);
+
+    log::info!("Changing current dir because else clippy won't find anything even if there is something (bug in --package argument?)");
+    let current_dir = current_dir().unwrap();
+    let new_dir = current_dir.parent().unwrap();
+    log::info!("New dir: {:#?}", new_dir);
+    set_current_dir(new_dir).unwrap();
 
     let metadata = cargo_metadata::MetadataCommand::new()
         .no_deps()
@@ -62,7 +72,13 @@ pub fn write_template(
     yml
 }
 
-pub fn write_tests(yml: &mut File, whitespace: &str, package: &str, fmt_and_fix: bool) {
+pub fn write_tests(
+    yml: &mut File,
+    whitespace: &str,
+    package: &str,
+    fmt_and_fix: bool,
+    package_name_currently_executing: &str,
+) {
     log::info!("Writing tests for package: {}", package);
 
     writeln!(yml, "{}- name: Build {}", whitespace, package).unwrap();
@@ -109,61 +125,52 @@ pub fn write_tests(yml: &mut File, whitespace: &str, package: &str, fmt_and_fix:
             "test_keyspace_for_testing".to_string(),
         );
 
-        // Format and fix project directly
-        log::info!("Cleaning...");
-        execute_cargo_command("clean", package, None, &env);
-        log::info!("Building...");
-        execute_cargo_command("build", package, None, &env);
+        // Execute this before running a command, because there are bugs preventing cargo to run
+        // checks the second time without a rebuild/touch (https://stackoverflow.com/questions/31125226/is-it-possible-to-have-cargo-always-show-warnings)
+        let clean_and_build = || {
+            // Format and fix project directly
+            // Only clean if the current package isn't the CI package, since that isn't possible
+            // (removing the executable while executing the program)
+            if package != package_name_currently_executing {
+                log::info!("Cleaning...");
+                execute_command("clean", package, None, &env);
+                log::info!("Building...");
+                execute_command("build", package, None, &env);
+            }
+        };
+
+        clean_and_build();
         log::info!("Formatting...");
-        execute_cargo_command("fmt", package, None, &env);
+        execute_command("fmt", package, None, &env);
+
+        clean_and_build();
         log::info!("Fixing...");
-        execute_cargo_command(
+        execute_command(
             "fix",
             package,
             Some(vec!["--all-features", "--allow-dirty"]),
             &env,
         );
+        // No clean and build needed for this
         log::info!("Testing...");
-        execute_cargo_command(
+        execute_command(
             "test",
             package,
+            // Test threads is 1, because in tests sometimes tables be added, keyspace are being recreated,
+            // if that all goes down at the same time, error will be thrown
             Some(vec!["--verbose", "--", "--test-threads=1"]),
             &env,
         );
 
+        clean_and_build();
+        log::info!("Running clippy...");
+        execute_command("clippy", package, Some(vec!["--", "-D", "warnings"]), &env);
+
         log::info!("Done verifying package");
     }
-
-    // TODO: This does not work yet
-    // Command::new("cargo")
-    //     .env("TEST_CDRS_DB_KEYSPACE_KEY", "test_keyspace_for_testing")
-    //     .args(&["build".to_string(), "--package".to_string(), format!("{}", package)])
-    //     .output()
-    //     .unwrap();
-    //
-    // // Use clippy to check for any errors
-    // let clippy_out = Command::new("cargo")
-    //     .env("TEST_CDRS_DB_KEYSPACE_KEY", "test_keyspace_for_testing")
-    //     .args(&[
-    //         "+nightly".to_string(),
-    //         "clippy".to_string(),
-    //         "--package".to_string(),
-    //         format!("{}", package),
-    //         "--".to_string(),
-    //         "-D".to_string(),
-    //         "warnings".to_string(),
-    //     ])
-    //     .output()
-    //     .unwrap();
-    //
-    // if !clippy_out.stderr.is_empty() && !clippy_out.status.success() {
-    //     let error = String::from_utf8(clippy_out.stderr).unwrap();
-    //
-    //     panic!("{}", error);
-    // }
 }
 
-pub fn execute_cargo_command(
+pub fn execute_command(
     command: &str,
     package: &str,
     extra_command: Option<Vec<&str>>,
@@ -192,7 +199,7 @@ pub fn execute_cargo_command(
         .unwrap();
 
     if !output.stderr.is_empty() && !output.status.success() {
-        panic!("{:#?}", String::from_utf8(output.stderr).unwrap());
+        panic!("{}", String::from_utf8(output.stderr).unwrap());
     }
 
     assert!(output.status.success());
